fix: 🔨 mergin auth store login and register
